Fix GoogleDrive save issue

+++ backend/src/main/java/com/languagematters/tessta/library/model/UserFile.java
@@ -1,24 +1,24 @@
 package com.languagematters.tessta.library.model;
 
 public class UserFile {
-    private int id;
-    private int userId;
+    private long id;
+    private long userId;
     private String name;
     private String originalFileName;
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
